usiu-chatbot/backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
# audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
gradio==5.24.0
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-google-genai==2.1.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.30
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.72.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
ruff==0.11.5
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.19.0
zstandard==0.23.0
protobuf>=6.30.0,<7.0.0
faiss-cpu
chromadb